Move create async resource bundle to lib
