+++ client-v2/src/bundles/frontsConfigBundle.ts
@@ -1,4 +1,4 @@
-import createAsyncResourceBundle, { State } from 'shared/util/createAsyncResourceBundle';
+import createAsyncResourceBundle, { State } from 'lib/createAsyncResourceBundle';
 import { FrontsConfig } from 'types/FaciaApi';
 
 export const {
