+++ client-v2/src/lib/createAsyncResourceBundle/__tests__/createAsyncResourceBundle.spec.ts
@@ -1,4 +1,4 @@
-import createAsyncResourceBundle from '../createAsyncResourceBundle';
+import createAsyncResourceBundle from '../';
 
 const { actions, reducer, selectors, initialState } = createAsyncResourceBundle(
   'books'
