+++ client-v2/src/lib/createAsyncResourceBundle/index.ts
